Finished refactor of discord functions and classes. Separated command cog into system, generation and document cogs. Added generate document contex command back to main discord function.
